Pull some of the generators in BagVerifiableTest into a shared trait
